requirements.txt
requests~=2.32.3
pytest~=8.2.2
setuptools~=68.2.0
openpyxl~=3.1.4
python-dotenv~=1.0.1